more generic solution for applying filters, separate directory for implementation
