+++ .gitignore
@@ -172,3 +172,4 @@ cython_debug/
 
 edge_detection/input
 edge_detection/output
+.idea
