+++ edge_detection/main.py
@@ -1,0 +1,4 @@
+from edge_detection import edge_detection
+
+if __name__ == "__main__":
+    edge_detection.run()
